requirements.txt
asgiref==3.9.1
autopep8==2.3.2
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pycodestyle==2.14.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
requests==2.32.3
